+++ CHANGELOG.md
@@ -5,7 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.5.0
+
+## Unrealeased
+
+### New Features
+
+* VolumeModel: to_geo (generate .geo file), to_msh (generate .msh file)
+* Gmsh (read .msh file)
+
+
+### Fixed
+
+* 
+
+
+### Performance improvements
+
+*
+
+
+### Refactorings
+
+* Define a Parent class 'Triangle' for Triangle2D/3D
+* 
+* 
+
+
+## Unrealeased
+
 
 ### New Features